# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PersonRole = typeof PersonRole[keyof typeof PersonRole];


export const PersonRole = {
  lead: 'lead',
  banker: 'banker',
} as const;

export interface Person {
  id: number;
  name: string;
  initials: string;
  role: PersonRole;
  /** Hex color used to identify the person on the calendar */
  color: string;
}

export type PersonInputRole = typeof PersonInputRole[keyof typeof PersonInputRole];


export const PersonInputRole = {
  lead: 'lead',
  banker: 'banker',
} as const;

export interface PersonInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  initials: string;
  role: PersonInputRole;
  color?: string;
}

export type PersonUpdateRole = typeof PersonUpdateRole[keyof typeof PersonUpdateRole];


export const PersonUpdateRole = {
  lead: 'lead',
  banker: 'banker',
} as const;

export interface PersonUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  initials?: string;
  role?: PersonUpdateRole;
  color?: string;
}

export type HolidayType = typeof HolidayType[keyof typeof HolidayType];


export const HolidayType = {
  annual: 'annual',
  sick: 'sick',
  public: 'public',
  other: 'other',
} as const;

export interface Holiday {
  id: number;
  personId: number;
  /** Inclusive start date (YYYY-MM-DD) */
  startDate: string;
  /** Inclusive end date (YYYY-MM-DD) */
  endDate: string;
  type: HolidayType;
  /** @nullable */
  notes?: string | null;
}

export type HolidayInputType = typeof HolidayInputType[keyof typeof HolidayInputType];


export const HolidayInputType = {
  annual: 'annual',
  sick: 'sick',
  public: 'public',
  other: 'other',
} as const;

export interface HolidayInput {
  personId: number;
  startDate: string;
  endDate: string;
  type?: HolidayInputType;
  notes?: string;
}

export type HolidayUpdateType = typeof HolidayUpdateType[keyof typeof HolidayUpdateType];


export const HolidayUpdateType = {
  annual: 'annual',
  sick: 'sick',
  public: 'public',
  other: 'other',
} as const;

export interface HolidayUpdate {
  personId?: number;
  startDate?: string;
  endDate?: string;
  type?: HolidayUpdateType;
  /** @nullable */
  notes?: string | null;
}

export type HolidayImportRowType = typeof HolidayImportRowType[keyof typeof HolidayImportRowType];


export const HolidayImportRowType = {
  annual: 'annual',
  sick: 'sick',
  public: 'public',
  other: 'other',
} as const;

export interface HolidayImportRow {
  /** Person name or initials to match against the team */
  name: string;
  startDate: string;
  endDate: string;
  type?: HolidayImportRowType;
  notes?: string;
}

export interface HolidayImportInput {
  rows: HolidayImportRow[];
}

export interface ImportResult {
  imported: number;
  skipped: number;
  unmatched: string[];
}

export interface CoverageDay {
  date: string;
  awayCount: number;
  overThreshold: boolean;
  personIds: number[];
}

export interface PersonAwayStat {
  personId: number;
  name: string;
  initials: string;
  daysOff: number;
}

export interface Summary {
  totalPeople: number;
  awayToday: number;
  awayThisWeek: number;
  upcomingBookings: number;
  /** Number of upcoming days (next 90) over the away threshold */
  conflictDays: number;
  perPerson: PersonAwayStat[];
}

export interface Settings {
  id: number;
  /** Maximum people allowed away before a conflict is flagged */
  maxAway: number;
  teamName: string;
}

export interface SettingsUpdate {
  maxAway?: number;
  teamName?: string;
}

export type ListHolidaysParams = {
/**
 * Inclusive start date (YYYY-MM-DD)
 */
from?: string;
/**
 * Inclusive end date (YYYY-MM-DD)
 */
to?: string;
personId?: number;
};

export type GetCoverageParams = {
/**
 * Inclusive start date (YYYY-MM-DD)
 */
from: string;
/**
 * Inclusive end date (YYYY-MM-DD)
 */
to: string;
};