# lib/api-zod/src/generated/types/holidayImportRow.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HolidayImportRowType } from './holidayImportRowType';

export interface HolidayImportRow {
  /** Person name or initials to match against the team */
  name: string;
  startDate: string;
  endDate: string;
  type?: HolidayImportRowType;
  notes?: string;
}