# lib/api-zod/src/generated/types/person.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PersonRole } from './personRole';

export interface Person {
  id: number;
  name: string;
  initials: string;
  role: PersonRole;
  /** Hex color used to identify the person on the calendar */
  color: string;
}