# lib/api-zod/src/generated/types/personAwayStat.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PersonAwayStat {
  personId: number;
  name: string;
  initials: string;
  daysOff: number;
}