# lib/api-zod/src/generated/types/personUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PersonUpdateRole } from './personUpdateRole';

export interface PersonUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  initials?: string;
  role?: PersonUpdateRole;
  color?: string;
}