# lib/api-zod/src/generated/types/settings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Settings {
  id: number;
  /** Maximum people allowed away before a conflict is flagged */
  maxAway: number;
  teamName: string;
}